chore: update lint and clean command
