+++ .eslintignore
@@ -1,2 +1,3 @@
 *.d.ts
 */dist
+*.test.ts
