+++ package.json
@@ -42,6 +42,7 @@
   },
   "pnpm": {
     "overrides": {
+      "typescript": "5.1.6",
       "axios@^0.x.x": "^0.27.2",
       "json5@<=1.0.1": "^1.0.2",
       "semver@<=4.3.2": "^4.3.6"
