+++ packages/swapkit/swapkit-sdk/package.json
@@ -52,7 +52,7 @@
   "repository": "https://github.com/thorswap/SwapKit.git",
   "scripts": {
     "build": "echo 'Build swapkit-sdk 🛠';NODE_OPTIONS=--max_old_space_size=32768 vite build && echo 'Build succeeded swapkit-sdk ✅'",
-    "clean": "rm -rf dist .turbo",
+    "clean": "rm -rf dist .turbo node_modules",
     "lint": "eslint ./ --ext .ts,.tsx --fix; tsc --noEmit",
     "test": "echo 'vitest --run'",
     "test:coverage": "echo 'vitest run --coverage'"
