+++ packages/wallets/evm-web3-wallets/package.json
@@ -44,7 +44,7 @@
   "repository": "https://github.com/thorswap/SwapKit.git",
   "scripts": {
     "build": "echo 'Build evm-web3-wallets 🛠'; vite build && echo 'Build succeeded evm-web3-wallets ✅'",
-    "clean": "rm -rf dist .turbo",
+    "clean": "rm -rf dist .turbo node_modules",
     "lint": "eslint ./ --ext .ts,.tsx --fix; tsc --noEmit",
     "test": "echo 'vitest --run'",
     "test:coverage": "echo 'vitest run --coverage'"
